Fixes #61 Simplify parent/child relationships
