+++ compiler/taglibs/Taglib/ImportedVariable.js
@@ -1,0 +1,8 @@
+var makeClass = require('raptor-util').makeClass;
+
+module.exports = makeClass({
+    $init: function() {
+        this.targetProperty = null;
+        this.expression = null;
+    }
+});
