+++ compiler/taglibs/Taglib/NestedVariable.js
@@ -1,0 +1,7 @@
+var makeClass = require('raptor-util').makeClass;
+
+module.exports = makeClass({
+    $init: function() {
+        this.name = null;
+    }
+});
