+++ compiler/taglibs/Taglib/Transformer.js
@@ -1,0 +1,42 @@
+var makeClass = require('raptor-util').makeClass;
+
+var nextTransformerId = 0;
+
+module.exports = makeClass({
+    $init: function() {
+        this.id = nextTransformerId++;
+        this.name = null;
+        this.tag = null;
+        this.path = null;
+        this.priority = null;
+        this._func = null;
+        this.properties = {};
+    },
+
+    getFunc: function () {
+        if (!this.path) {
+            throw new Error('Transformer path not defined for tag transformer (tag=' + this.tag + ')');
+        }
+
+        if (!this._func) {
+            var transformer = require(this.path);
+
+            if (typeof transformer === 'function') {
+                if (transformer.prototype.process) {
+                    var Clazz = transformer;
+                    var instance = new Clazz();
+                    instance.id = this.id;
+                    this._func = instance.process.bind(instance);
+                } else {
+                    this._func = transformer;
+                }
+            } else {
+                this._func = transformer.process || transformer.transform;
+            }
+        }
+        return this._func;
+    },
+    toString: function () {
+        return '[Taglib.Transformer: ' + this.path + ']';
+    }
+});
