+++ compiler/taglibs/taglib-loader/loader.js
@@ -1,0 +1,3 @@
+exports.taglibLoader = require('./loader-taglib');
+exports.tagLoader = require('./loader-tag');
+exports.attributeLoader = require('./loader-attribute');
