+++ compiler/taglibs/taglib-loader/package.json
@@ -1,0 +1,6 @@
+{
+    "browser": {
+        "./scanTagsDir.js": "./scanTagsDir-browser.js",
+        "./taglib-reader.js": "./taglib-reader-browser.js"
+    }
+}
