+++ compiler/taglibs/taglib-loader/scanTagsDir-browser.js
@@ -1,0 +1,3 @@
+module.exports = function scanTagsDir() {
+    // no-op in the browser
+};
