+++ compiler/taglibs/taglib-loader/scanTagsDir.js
@@ -1,0 +1,128 @@
+var nodePath = require('path');
+var fs = require('fs');
+var jsonminify = require('jsonminify');
+var tagDefFromCode = require('./tag-def-from-code');
+var loader = require('./loader');
+
+function createDefaultTagDef() {
+    return {
+            attributes: {
+                '*': {
+                    type: 'string',
+                    targetProperty: null,
+                    preserveName: false
+                }
+            }
+         };
+}
+
+
+/**
+ * @param {String} tagsConfigPath path to tag definition file
+ * @param {String} tagsConfigDirname path to directory of tags config file (should be path.dirname(tagsConfigPath))
+ * @param {String|Object} dir the path to directory to scan
+ * @param {String} taglib the taglib that is being loaded
+ */
+module.exports = function scanTagsDir(tagsConfigPath, tagsConfigDirname, dir, taglib) {
+    var prefix;
+
+    if (typeof dir === 'object') {
+        prefix = dir.prefix;
+        dir = dir.path;
+    }
+
+    if (prefix == null) {
+        // no prefix by default
+        prefix = '';
+    }
+
+    dir = nodePath.resolve(tagsConfigDirname, dir);
+    var children = fs.readdirSync(dir);
+    var rendererJSFile;
+
+    for (var i=0, len=children.length; i<len; i++) {
+        rendererJSFile = null;
+        var childFilename = children[i];
+        if (childFilename === 'node_modules') {
+            continue;
+        }
+
+        var tagName = prefix + childFilename;
+        var tagDirname = nodePath.join(dir, childFilename);
+        var tagFile = nodePath.join(dir, childFilename, 'marko-tag.json');
+        var tag = null;
+        var rendererFile = nodePath.join(dir, childFilename, 'renderer.js');
+        var indexFile = nodePath.join(dir, childFilename, 'index.js');
+        var templateFile = nodePath.join(dir, childFilename, 'template.marko');
+        var tagDef = null;
+
+        // Record dependencies so that we can check if a template is up-to-date
+        taglib.addInputFile(tagFile);
+        taglib.addInputFile(rendererFile);
+
+        if (fs.existsSync(tagFile)) {
+            // marko-tag.json exists in the directory, use that as the tag definition
+            tagDef = JSON.parse(jsonminify(fs.readFileSync(tagFile, {encoding: 'utf8'})));
+            if (!tagDef.renderer && !tagDef.template) {
+                if (fs.existsSync(rendererFile)) {
+                    tagDef.renderer = rendererFile;
+                } else if (fs.existsSync(indexFile)) {
+                    tagDef.renderer = indexFile;
+                } else if (fs.existsSync(templateFile)) {
+                    tagDef.template = templateFile;
+                } else if (fs.existsSync(templateFile + ".html")) {
+                    tagDef.template = templateFile + ".html";
+                } else {
+                    throw new Error('Invalid tag file: ' + tagFile + '. Neither a renderer or a template was found for tag.');
+                }
+            }
+
+            tag = loader.tagLoader.loadTag(tagDef, tagsConfigPath, taglib, tagDirname);
+            tag.name = tag.name || tagName;
+            taglib.addTag(tag);
+        } else {
+            // marko-tag.json does *not* exist... checking for a 'renderer.js'
+            if (fs.existsSync(rendererFile)) {
+                rendererJSFile = rendererFile;
+            } else if (fs.existsSync(indexFile)) {
+                rendererJSFile = indexFile;
+            } else {
+                var exTemplateFile;
+                if (fs.existsSync(templateFile)) {
+                    exTemplateFile = templateFile;
+                }
+                else if (fs.existsSync(templateFile + ".html")){
+                    exTemplateFile = templateFile + ".html";
+                }
+                if(exTemplateFile){
+                    var templateCode = fs.readFileSync(exTemplateFile, {encoding: 'utf8'});
+                    tagDef = tagDefFromCode.extractTagDef(templateCode);
+                    if (!tagDef) {
+                        tagDef = createDefaultTagDef();
+                    }
+
+                    tagDef.template = exTemplateFile;
+                }
+            }
+
+            if (rendererJSFile) {
+                var rendererCode = fs.readFileSync(rendererJSFile, {encoding: 'utf8'});
+                tagDef = tagDefFromCode.extractTagDef(rendererCode);
+                if (!tagDef) {
+                     tagDef = createDefaultTagDef();
+                }
+
+                tagDef.renderer  = rendererJSFile;
+                tag = loader.tagLoader.loadTag(tagDef, tagsConfigPath, taglib, tagDirname);
+                tag.name = tagName;
+                taglib.addTag(tag);
+            }
+
+            if (tagDef) {
+                tag = loader.tagLoader.loadTag(tagDef, tagsConfigPath, taglib, tagDirname);
+                tag.name = tagName;
+                taglib.addTag(tag);
+            }
+        }
+    }
+};
