+++ compiler/taglibs/taglib-loader/taglib-reader-browser.js
@@ -1,0 +1,11 @@
+exports.readTaglib = function (path) {
+    var taglibProps;
+
+    try {
+        taglibProps = require(path);
+    } catch(e) {
+        throw new Error('Unable to parse taglib JSON at path "' + path + '". Exception: ' + e);
+    }
+
+    return taglibProps;
+};
