+++ compiler/taglibs/taglib-loader/taglib-reader.js
@@ -1,0 +1,13 @@
+var fs = require('fs');
+var jsonminify = require('jsonminify');
+
+exports.readTaglib = function (path) {
+    var json = fs.readFileSync(path, 'utf8');
+
+    try {
+        var taglibProps = JSON.parse(jsonminify(json));
+        return taglibProps;
+    } catch(e) {
+        throw new Error('Unable to parse taglib at path "' + path + '". Error: ' + e);
+    }
+};
