+++ test/fixtures/taglib-shorthand/marko-taglib.json
@@ -1,0 +1,25 @@
+{
+    "<shorthand-checkbox>": {
+        "@label <label>": "string",
+        "@checked": "boolean",
+        "<checked>": "boolean"
+    },
+    "<shorthand-overlay>": {
+        "nested-tags": {
+            "body": {
+                "@condensed": "boolean"
+            }
+        }
+    },
+    "tags": {
+        "shorthand-button": {
+            "@label": "string"
+        },
+        "shorthand-tabs": {
+            "@tabs <tab>[]": {
+                "@label": "string"
+            },
+            "@orientation": "string"
+        }
+    }
+}
