+++ test/fixtures/taglib/test-nested-tags-overlay/renderer.js
@@ -1,0 +1,15 @@
+var marko = require('../../../../');
+var template = marko.load(require.resolve('./template.marko'));
+
+exports.renderer = function(input, out) {
+    var header = input.header;
+    var body = input.body;
+    var footer = input.footer;
+
+    template.render({
+        header: header,
+        body: body,
+        footer: footer
+    }, out);
+
+};
