+++ test/fixtures/taglib/test-nested-tags-overlay/template.marko
@@ -1,0 +1,13 @@
+<div class="overlay">
+    <div class="overlay-header ${data.header.className}" if="data.header">
+        <invoke function="data.header.renderBody(out)"/>
+    </div>
+
+    <div class="overlay-body ${data.body.className}" if="data.body">
+        <invoke function="data.body.renderBody(out)"/>
+    </div>
+
+    <div class="overlay-footer ${data.footer.className}" if="data.footer">
+        <invoke function="data.footer.renderBody(out)"/>
+    </div>
+</div>
