+++ test/fixtures/taglib/test-nested-tags-tabs/marko-tag.json
@@ -1,0 +1,7 @@
+{
+    "renderer": "./renderer",
+    "@orientation": "string",
+    "@tabs <tab>[]": {
+        "@title": "string"
+    }
+}
