+++ test/fixtures/templates/nested-tags-repeated/expected.html
@@ -1,0 +1,1 @@
+<div class="tabs"><ul><li>Tab 1</li><li>Tab 3</li></ul><div class="tab-content"><div class="tab">Tab 1 content</div><div class="tab">Tab 3 content</div></div></div>
