+++ test/fixtures/templates/nested-tags-repeated/template.marko
@@ -1,0 +1,14 @@
+<test-nested-tags-tabs>
+    <test-nested-tags-tabs.tab title="Tab 1">
+
+        Tab 1 content
+    </test-nested-tags-tabs.tab>
+
+    <test-nested-tags-tabs.tab title="Tab 2" if="false">
+        Tab 2 content
+    </test-nested-tags-tabs.tab>
+
+    <test-nested-tags-tabs.tab title="Tab 3">
+        Tab 3 content
+    </test-nested-tags-tabs.tab>
+</test-nested-tags-tabs>
