+++ test/fixtures/templates/nested-tags-repeated/test.js
@@ -1,0 +1,1 @@
+exports.templateData = {};
