+++ test/fixtures/templates/nested-tags/expected.html
@@ -1,0 +1,1 @@
+<div class="overlay"><div class="overlay-header ">Header content!</div><div class="overlay-body my-body">Body content</div><div class="overlay-footer my-footer">Footer content</div></div>
