+++ test/fixtures/templates/nested-tags/template.marko
@@ -1,0 +1,10 @@
+<test-nested-tags-overlay header="data.header">
+
+    <test-nested-tags-overlay.body class="my-body">
+        Body content
+    </test-nested-tags-overlay.body>
+
+    <test-nested-tags-overlay.footer class="my-footer">
+        Footer content
+    </test-nested-tags-overlay.footer>
+</test-nested-tags-overlay>
