+++ test/fixtures/templates/nested-tags/test.js
@@ -1,0 +1,7 @@
+exports.templateData = {
+    header: {
+        renderBody: function(out) {
+            out.write('Header content!');
+        }
+    }
+};
